requirements.txt
streamlit>=1.28.0
fast-graphrag>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.15.0
networkx>=3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
markdown>=3.4.0
streamlit-option-menu>=0.3.6
streamlit-ace>=0.1.1
streamlit-aggrid>=0.3.4
